refactor: paging에서 entity와 dto 정리 - entity > Article.java : 보안상 고려해야 하는 것이므로 엔티티로 관리
- dto
> ArticleCursorResp.java : cursor의 resp
> ArticleOffsetResp.java : offset의 resp
> ArticleList.java : resp로 만들 articleList의 형태
> ArticleReq.java : req의 형태
> ArticleResp.java : resp의 형태 (ArticleCursorResp, ArticleOffsetResp,ArticleList에서 사용됨)
