+++ src/main/java/com/sandbox/domain/articles/dao/ArticleDao.java
@@ -1,6 +1,5 @@
 package com.sandbox.domain.articles.dao;
 
-import com.sandbox.domain.articles.dto.Article;
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.PersistenceContext;
 import lombok.RequiredArgsConstructor;
