+++ src/main/java/com/sandbox/domain/articles/dto/ArticleList.java
@@ -2,13 +2,11 @@ package com.sandbox.domain.articles.dto;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import lombok.Setter;
 
 import java.util.List;
 
-@Getter @Setter
+@Getter
 @NoArgsConstructor
 public class ArticleList {
-    private List<Article> articles;
-
+    private List<ArticleResp> articles;
 }
