+++ src/main/java/com/sandbox/domain/articles/dto/ArticleResp.java
@@ -3,14 +3,12 @@ package com.sandbox.domain.articles.dto;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import lombok.Setter;
 
-import java.util.List;
-
-@Getter @Setter
+@Getter
 @NoArgsConstructor
 @AllArgsConstructor
-public class ArticleCursorResp {
-    private Integer lastId;
-    private List<ArticleResp> articles;
+public class ArticleResp {
+    private int id;
+    private String title;
+    private String createdAt;
 }
