+++ src/main/java/com/sandbox/domain/articles/entity/Article.java
@@ -1,21 +1,19 @@
-package com.sandbox.domain.articles.dto;
+package com.sandbox.domain.articles.entity;
 
 import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-import java.time.LocalDateTime;
-
 @Entity
 @Getter @Setter
 @NoArgsConstructor
+@AllArgsConstructor
 public class Article {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private int id;
-
     private String title;
-
-    private String createdAt; //지금 샌드박스에 있는 시간 형식!
+    private String createdAt;
 }
