+++ src/main/java/com/sandbox/domain/articles/service/ArticleService.java
@@ -1,7 +1,5 @@
 package com.sandbox.domain.articles.service;
 
-import com.sandbox.domain.articles.dto.Article;
-
 import com.sandbox.domain.articles.dto.ArticleCursorResp;
 import com.sandbox.domain.articles.dto.ArticleOffsetResp;
