+++ src/main/java/com/sandbox/domain/articles/service/ArticleServiceImpl.java
@@ -1,7 +1,6 @@
 package com.sandbox.domain.articles.service;
 
 import com.sandbox.domain.articles.dao.ArticleDao;
-import com.sandbox.domain.articles.dto.Article;
 import com.sandbox.domain.articles.dto.ArticleCursorResp;
 import com.sandbox.domain.articles.dto.ArticleOffsetResp;
 import lombok.RequiredArgsConstructor;
